pyproject.toml
[tool.poetry]
name = "detect-missing-init-files"
version = "0.1.0"
description = ""
authors = ["Luuk Verweij <luuk_verweij@msn.com>"]

packages = [
    { include="hook", from="." },
]

[tool.poetry.dependencies]
python = "^3.8"
GitPython = "^3.1.24"

[tool.poetry.dev-dependencies]
pyfakefs = "^4.5.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
detect-missing-init = "hook.detect_missing_init:main"
